chore: updated file structure to be more organised
