+++ packages/pieces/apps/src/lib/clickup/actions/tasks/create-task.ts
@@ -1,5 +1,5 @@
 import { createAction, Property, HttpMethod, getAccessTokenOrThrow } from "@activepieces/framework";
-import { clickupCommon, callClickUpApi } from "../common";
+import { clickupCommon, callClickUpApi } from "../../common";
 
 
 export const createClickupTask = createAction({
