+++ packages/pieces/apps/src/lib/clickup/index.ts
@@ -1,10 +1,10 @@
 import { createPiece } from "@activepieces/framework";
 import { getClickupTaskCommants } from "./actions/comments/get-task-comments";
-import { createClickupFolderlessList } from "./actions/create-folderless-list";
-import { createClickupTask } from "./actions/create-task";
-import { getClickupList } from "./actions/get-list";
-import { getClickupSpace } from "./actions/get-space";
-import { getClickupSpaces } from "./actions/get-spaces";
+import { createClickupFolderlessList } from "./actions/lists/create-folderless-list";
+import { createClickupTask } from "./actions/tasks/create-task";
+import { getClickupList } from "./actions/lists/get-list";
+import { getClickupSpace } from "./actions/spaces/get-space";
+import { getClickupSpaces } from "./actions/spaces/get-spaces";
 
 export const clickup = createPiece({
     name: 'clickup',
